Move entire content of old folder
Instead of moving individual *.java files to the root of the new folder, we could instead move the entire content of the old folder to the new path.  

I'm not versed in android native development, I'm not sure if this would cause issues with other projects, but in React Native, there's a new folder: newarchitecure, which if for, well, pretty self explanatory. It contains 3 folders, each containing 1 file.